update: move subcomponent to own subdirectory
