+++ src/views/pages/index.ts
@@ -2,3 +2,4 @@ export * from './ErrorPage'
 export * from './GamePage'
 export * from './SearchPage'
 export * from './HomePage'
+export { SearchControls } from '@views/pages/SearchControls/SearchControls.tsx'
